In Selection Works... and Match Case just stopped Working
This works well, but now the function is disregarding the Match Case checkbox.   If the Match Case checkbox is active, only characters of the same case should be replaced, however if the Match Case checkbox is NOT active all instances of find_term_entry should be replace regardless of case.  Currently, the function is only matching the case as if the Match Case checkbox is active all the time.