Rewritting find Function
I think the problem is that the find function is searching for one instance at a time, and it stops when that first instance doesn't meet the criteria of the Regular Express.  So, I'm re-writing the find function into a Find All function, then running the RE over an array of those results.  That way, the code doesn't grind to a halt.  ... but its that really the most efficient method.  I need to find an example of this function in the wild.