Made opening the navigation_frame look more like the closing
It's almost perfect.  I only need to change one thing. When I press the close button, the navigation_frame slides closed smoothly as 
the background colors of the bottom_border_frame and the top_border_frame change.  HOWEVER, when I click the navigation_button, the background colors of the 
bottom_border_frame and the top_border_frame change almost instantly.